Support opening PDFs larger than 2 GiB (#1293)
* Add tests for large PDF handling and improve file reading

- Introduced a new test in `LocalTests.cs` for creating and reading PDFs larger than two gigabytes using `LargePdfTestDocumentGenerator`.
- Updated `FileHeaderOffset.cs` to change the `Value` type from `int` to `long` for better handling of large values.
- Modified `PdfDocumentFactory.cs` to use `FileStream` for opening PDFs, enhancing memory efficiency.
- Added `PdfDocumentOpenTests` class to verify opening of large PDFs with a sparse XRef offset.

* Add tests for handling large PDF files

Introduced `RealBigFileTests` to test creation and reading of PDFs larger than 2 GiB. Updated `IntegrationDocumentTests` with a new test for opening large sparse documents and added utility methods for document creation. Removed `LocalTests` and `PdfDocumentOpenTests` to refactor the test structure, focusing on integration tests for large files.

* Revert Open(string) to File.ReadAllBytes; route large-file tests through Open(Stream)

The previous change to PdfDocumentFactory.Open(string) switched it from
File.ReadAllBytes to a FileStream wrapped in StreamInputBytes(shouldDispose: true),
so the document owned a live file handle for its lifetime. That is a silent
breaking change: Open(string) has never required disposal, and existing callers
(many of which don't wrap the document in a using block) would suddenly be
holding the file open and blocking re-open or delete until GC ran. Only
Open(Stream) is documented to require the caller to manage the stream.

This restores the pre-existing byte-array path for Open(string). Files larger
than 2 GiB cannot be loaded that way (byte[] is int.MaxValue-capped), which is
the same constraint that has always existed: callers with >2 GiB files must use
Open(Stream) with their own FileStream, and the two large-file tests are
updated to do exactly that. FileHeaderOffset.Value stays long since it's
internal-only and removes a lossy int cast.

* Drop optional real >2 GiB PDF repro test per review

Per PR #1293 reviewer feedback, removing RealBigFileTests and its
LargePdfTestDocumentGenerator helper. The actual regression coverage
lives in IntegrationDocumentTests.CanOpenLargeSparseDocumentAndReadPage,
which uses a sparse file with a >2 GiB xref offset.

---------

+++ src/UglyToad.PdfPig/Parser/PdfDocumentFactory.cs
@@ -127,7 +127,7 @@
 
             var version = FileHeaderParser.Parse(scanner, inputBytes, parsingOptions.UseLenientParsing, parsingOptions.Logger);
 
-            var fileHeaderOffset = new FileHeaderOffset((int)version.OffsetInFile);
+            var fileHeaderOffset = new FileHeaderOffset(version.OffsetInFile);
 
             var initialParse = FirstPassParser.Parse(
                 fileHeaderOffset,
